Create code_maker.sh
Bash script for the first two steps of running the maker2 annotation software. Note that albeit working, this did not work well for me. I am pretty sure this is the script that was used to generate the annotation I in the end used. But I have 4 versions on my uppmax so contact me if interested!