Create code_phylobayes.sh
Bash script to run phylobayes using extra statistical analysis. Already used in other smaller round/redundant code commented out, i.e this script is not polished or that nice but it runs. I have an additional one using multiple chains, but this did not finish in time therefore is not used in the results nor analysis part, but it is a working phylobayes script for multiple chains available upon request.